Example dataset for use in OBIC package

This table contains a series of agricultural fields with soil properties needed for illustration OBIC.

binnenveld

Format

A data.frame with 65 rows and 55 columns:

ID

A field id (numeric)

YEAR

The year that the crop is grown (integer)

B_LU_BRP

A series with crop codes given the crop rotation plan (integer, source: the BRP)

B_SC_WENR

The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006) (character).

B_GWL_CLASS

The groundwater table class (character)

B_SOILTYPE_AGR

The agricultural type of soil (character)

B_HELP_WENR

The soil type abbreviation, derived from 1:50.000 soil map (character)

B_AER_CBS

The agricultural economic region in the Netherlands (CBS, 2016) (character)

A_SOM_LOI

The percentage organic matter in the soil (%) (numeric)

A_CLAY_MI

The clay content of the soil (%) (numeric)

A_SAND_MI

The sand content of the soil (%) (numeric)

A_SILT_MI

The silt content of the soil (%) (numeric)

A_PH_CC

The acidity of the soil, measured in 0.01M CaCl2 (-) (numeric)

A_CACO3_IF

The carbonate content of the soil (%) (numeric)

A_N_RT

The organic nitrogen content of the soil in mg N / kg (numeric)

A_CN_FR

The carbon to nitrogen ratio (-) (numeric)

A_COM_FR

The carbon fraction of soil organic matter (%) (numeric)

A_S_RT

The total Sulfur content of the soil (in mg S per kg) (numeric)

A_N_PMN

The potentially mineralizable N pool (mg N / kg soil) (numeric)

A_P_AL

The P-AL content of the soil (numeric)

A_P_CC

The plant available P content, extracted with 0.01M CaCl2 (mg / kg) (numeric)

A_P_WA

The P-content of the soil extracted with water (mg P2O5 / 100 ml soil) (numeric)

A_CEC_CO

The cation exchange capacity of the soil (mmol+ / kg), analysed via Cobalt-hexamine extraction (numeric)

A_CA_CO_PO

The The occupation of the CEC with Ca (%) (numeric)

A_MG_CO_PO

The The occupation of the CEC with Mg (%) (numeric)

A_K_CO_PO

The occupation of the CEC with K (%) (numeric)

A_K_CC

The plant available K content, extracted with 0.01M CaCl2 (mg / kg) (numeric)

A_MG_CC

The plant available Mg content, extracted with 0.01M CaCl2 (ug / kg) (numeric)

A_MN_CC

The plant available Mn content, extracted with 0.01M CaCl2 (ug / kg) (numeric)

A_ZN_CC

The plant available Zn content, extracted with 0.01M CaCl2 (ug / kg) (numeric)

A_CU_CC

The plant available Cu content, extracted with 0.01M CaCl2 (ug / kg) (numeric)

A_EW_BCS

The presence of earth worms (optional, score 0-1-2, numeric)

A_SC_BCS

The presence of compaction of subsoil (optional, score 0-1-2, numeric)

A_GS_BCS

The presence of waterlogged conditions, gley spots (optional, score 0-1-2, numeric)

A_P_BCS

The presence / occurrence of water puddles on the land, ponding (optional, score 0-1-2, numeric)

A_C_BCS

The presence of visible cracks in the top layer (optional, score 0-1-2, numeric)

A_RT_BCS

The presence of visible tracks / rutting or trampling on the land (optional, score 0-1-2, numeric)

A_RD_BCS

The rooting depth (optional, score 0-1-2, numeric)

A_SS_BCS

The soil structure (optional, score 0-1-2, numeric)

A_CC_BCS

he crop cover on the surface (optional, score 0-1-2, numeric)

M_COMPOST

The frequency that compost is applied (optional, every x years, numeric)

M_GREEN

A soil measure. Are catch crops sown after main crop (optional, option: yes or no, boolean)

M_NONBARE

A soil measure. Is parcel for 80 percent of the year cultivated and 'green' (optional, option: yes or no, boolean)

M_EARLYCROP

A soil measure. Use of early crop varieties to avoid late harvesting (optional, option: yes or no, boolean)

M_SLEEPHOSE

A soil measure. Is sleephose used for slurry application (optional, option: yes or no, boolean)

M_DRAIN

A soil measure. Are under water drains installed in peaty soils (optional, option: yes or no, boolean)

M_DITCH

A soil measure. Are ditched maintained carefully and slib applied on the land (optional, option: yes or no, boolean)

M_UNDERSEED

A soil measure. Is grass used as second crop in between maize rows (optional, option: yes or no, boolean)

M_LIME

A soil measure. Has field been limed in last three years (option: yes or no, boolean)

M_NONINVTILL

A soil measure. Non inversion tillage (option: yes or no, boolean)

M_SSPM

A soil measure. Soil Structure Protection Measures, such as fixed driving lines, low pressure tires, and light weighted machinery (option: yes or no, boolean)

M_SOLIDMANURE

A soil measure. Use of solid manure (option: yes or no, boolean)

M_STRAWRESIDUE

A soil measure. Application of straw residues (option: yes or no, boolean)

M_MECHWEEDS

A soil measure. Use of mechanical weed protection (option: yes or no, boolean)

M_PESTICIDES_DST

A soil measure. Use of DST for pesticides (option: yes or no, boolean)