Add Makkink correction factors and crop cover to crop rotation table

This function adds Makkink correction factors for ET and crop cover to the crop rotation table

calc_makkink(B_LU_BRP)

Arguments

B_LU_BRP

(numeric) The crop code from the BRP

Value

A datatable with the crop dependent Makkink correction factor per month. Output is a single data.table with for each B_LU_BRP code the monthly correction factor. Columns of the data.table are: crop_makkink, month, year, mcf and crop_cover.

Examples

calc_makkink(B_LU_BRP = 265)
#>     crop_makkink month year  mcf crop_cover
#>  1:     grasland     1    1 0.80          1
#>  2:     grasland     2    1 0.90          1
#>  3:     grasland     3    1 0.95          1
#>  4:     grasland     4    1 1.20          1
#>  5:     grasland     5    1 1.20          1
#>  6:     grasland     6    1 1.20          1
#>  7:     grasland     7    1 1.10          1
#>  8:     grasland     8    1 1.10          1
#>  9:     grasland     9    1 1.10          1
#> 10:     grasland    10    1 1.00          1
#> 11:     grasland    11    1 0.90          1
#> 12:     grasland    12    1 0.80          1
calc_makkink(B_LU_BRP = c(265,1019))
#>     crop_makkink month year  mcf crop_cover
#>  1:     grasland     1    1 0.80          1
#>  2:     grasland     2    1 0.90          1
#>  3:     grasland     3    1 0.95          1
#>  4:     grasland     4    1 1.20          1
#>  5:     grasland     5    1 1.20          1
#>  6:     grasland     6    1 1.20          1
#>  7:     grasland     7    1 1.10          1
#>  8:     grasland     8    1 1.10          1
#>  9:     grasland     9    1 1.10          1
#> 10:     grasland    10    1 1.00          1
#> 11:     grasland    11    1 0.90          1
#> 12:     grasland    12    1 0.80          1
#> 13:       overig     1    2 0.95          1
#> 14:       overig     2    2 0.95          1
#> 15:       overig     3    2 0.95          1
#> 16:       overig     4    2 1.00          1
#> 17:       overig     5    2 1.00          1
#> 18:       overig     6    2 1.00          1
#> 19:       overig     7    2 1.00          1
#> 20:       overig     8    2 1.00          1
#> 21:       overig     9    2 1.00          1
#> 22:       overig    10    2 0.95          1
#> 23:       overig    11    2 0.95          1
#> 24:       overig    12    2 0.95          1
#>     crop_makkink month year  mcf crop_cover