Calculate nitrogen use efficiency and leaching based on N surplus

This function gives an indication of the nitrogen use efficiency, the function calculates the N surplus and the resulting N leaching

calc_n_efficiency(
  B_LU_BRP,
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  B_AER_CBS,
  A_SOM_LOI,
  A_CLAY_MI,
  D_PBI,
  D_K,
  D_PH_DELTA,
  D_NLV,
  M_GREEN = FALSE,
  B_FERT_NORM_FR = 1
)

Arguments

B_LU_BRP

(numeric) The crop code from the BRP

B_SOILTYPE_AGR

(character) The agricultural type of soilBRP

B_GWL_CLASS

(character) The groundwater table class

B_AER_CBS

(character) The agricultural economic region in the Netherlands (CBS, 2016)

A_SOM_LOI

(numeric) The percentage organic matter in the soil (%)

A_CLAY_MI

(numeric) The clay content of the soil (%)

D_PBI

(numeric) The value of phosphate availability calculated by calc_phosphate_availability

D_K

(numeric) The value of K-index calculated by calc_potassium_availability

D_PH_DELTA

(numeric) The pH difference with the optimal pH.

D_NLV

(numeric) The value of NLV calculated by calc_nlv

M_GREEN

(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)

B_FERT_NORM_FR

(numeric) The fraction of the application norm utilized

Value

The estimated index for the nitrogen use efficiency, as being affected by soil properties. A numeric value.

Examples

calc_n_efficiency(1019,'dekzand','GtIV','Zuidwest-Brabant',4.5,3.5,0.8,0.6,0.2,78,FALSE,1)
#> [1] 36.05
calc_n_efficiency(256,'veen','GtII','Centraal Veehouderijgebied',4.5,3.5,0.8,0.6,0.2,250,FALSE,1)
#> [1] 8.25