Calculate amount of organic carbon

This function calculates the amount of organic carbon in the soil

calc_organic_carbon(A_SOM_LOI, D_BDS, D_RD)

Arguments

A_SOM_LOI

(numeric) The percentage organic matter in the soil

D_BDS

(numeric) The bulk density of the soil

D_RD

(numeric) The root depth of the crop

Value

The total amount of Carbon in the soil (kg C / ha). A numeric value.

Examples

calc_organic_carbon(A_SOM_LOI = 4.3, D_BDS = 1100, D_RD = 0.2)
#> [1] 54868
calc_organic_carbon(A_SOM_LOI = c(1,4.3), D_BDS = c(1100,1300), D_RD = c(0.2,0.6))
#> [1]  12760 194532