Calculate the difference between pH and optimum

This functions calculates the difference between the measured pH and the optimal pH according to the Bemestingsadvies

calc_ph_delta(
  B_LU_BRP,
  B_SOILTYPE_AGR,
  A_SOM_LOI,
  A_CLAY_MI,
  A_PH_CC,
  D_CP_STARCH,
  D_CP_POTATO,
  D_CP_SUGARBEET,
  D_CP_GRASS,
  D_CP_MAIS,
  D_CP_OTHER
)

Arguments

B_LU_BRP

(numeric) The crop code from the BRP

B_SOILTYPE_AGR

(character) The agricultural type of soil

A_SOM_LOI

(numeric) The organic matter content of soil in percentage

A_CLAY_MI

(numeric) The percentage A_CLAY_MI present in the soil

A_PH_CC

(numeric) The pH-CaCl2 of the soil

D_CP_STARCH

(numeric) The fraction of starch potatoes in the crop plan

D_CP_POTATO

(numeric) The fraction of potatoes (excluding starch potatoes) in the crop plan

D_CP_SUGARBEET

(numeric) The fraction of sugar beets in the crop plan

D_CP_GRASS

(numeric) The fracgtion of grass in the crop plan

D_CP_MAIS

(numeric) The fraction of mais in the crop plan

D_CP_OTHER

(numeric) The fraction of other crops in the crop plan

References

Handboek Bodem en Bemesting tabel 5.1, 5.2 en 5.3