Calculate the indicator for SLV (deprecated)

This function calculates the indicator for the the S-index by using the SLV calculated by calc_slv

This function calculates the indicator for the the S-index by using the SLV calculated by calc_slv

ind_sulpher(D_SLV, B_LU_BRP, B_SOILTYPE_AGR, B_AER_CBS)

ind_sulpher(D_SLV, B_LU_BRP, B_SOILTYPE_AGR, B_AER_CBS)

Arguments

D_SLV

(numeric) The value of SLV calculated by calc_slv

B_LU_BRP

(numeric) The crop code (gewascode) from the BRP

B_SOILTYPE_AGR

(character) The type of soil

B_AER_CBS

(character) The agricultural economic region in the Netherlands (CBS, 2016)

Value

The evaluated score for the soil function to supply sulfur for crop uptake. A numeric value between 0 and 1.

The evaluated score for the soil function to supply sulfur for crop uptake. A numeric value between 0 and 1.

Details

Pl

Examples

ind_sulpher(D_SLV = 15,B_LU_BRP = 256,B_SOILTYPE_AGR = 'dekzand',
B_AER_CBS = 'Rivierengebied')
#> [1] 0.9963242
ind_sulpher(c(10,15,35),c(256,1019,1019),rep('rivierklei',3),rep('Rivierengebied',3))
#> [1] 0.3083304 0.6454846 0.9999589
 
ind_sulpher(D_SLV = 15,B_LU_BRP = 256,B_SOILTYPE_AGR = 'dekzand',B_AER_CBS = 'Rivierengebied')
#> [1] 0.9963242
ind_sulpher(c(10,15,35),c(256,1019,1019),rep('rivierklei',3),rep('Rivierengebied',3))
#> [1] 0.3083304 0.6454846 0.9999589