Damage and reproduction of soil-borne pathogens and pests on crops

This table includes information from aaltjesschema (April 2021), a website where information is collected on the vulnerability of crops to plant parasitic nematodes and diseases that use nematodes as vector.

nema.crop.rot.obic

Format

An object of class data.table (inherits from data.frame) with 7059 rows and 21 columns.

Details

crop

crop as called in aaltjesschema

name_scientific

scientific name of nematode

propagation

how easily a nematode can propagate on a crop given as strings with 5 classes

damage

strings indicating how much damage a nematode can inflict on a crop, with 5 classes

cultivar_dependent

boolean whether there are differences in propgation between cultivars of the crop

serotype_dependant

boolean whether there are differences in propagation between serotypes of the pathogen

dalgrond

boolean whether information is valid for soiltype 'dalgrond'

klei

boolean whether information is valid for soiltype 'klei'

loess

boolean whether information is valid for soiltype 'loess'

zand

boolean whether information is valid for soiltype 'zand'

zavel

boolean whether information is valid for soiltype 'zavel'

info

string whether there is information on propgation, differentiating between none, yes, and some

name_common

string, common name of pathogen in Dutch, if no common name is available, scientific name is given

nema_name

string, full name of pathogen in aaltjesschema, includes common and scientific name

grondsoort

string with letters indicating for which soil the information is valid

groen_br

boolean indicating that the crop is a green manure on fallow

groen_vs

boolean indicating that the crop is a green manure in early stubble

groen_od

boolean indicating that the crop is a green manure beneath cover crop

groen_ls

boolean indicating that the crop is a green manure in late stubble

groen_st

boolean indicating that the crop is a green manure as drifting deck

crop_name_scientific

string, scientific name of crop species or genus