Table with fractions of excess N which runs off to groundwater and surface water

This table contains the fractions of N overshot which runs off to groundwater / surface water, per soil type, crop type, and groundwater table

nleach_table

Format

An object of class data.table (inherits from data.frame) with 558 rows and 7 columns.

Details

gewas

crop type

bodem

soil type

ghg

Lower value for groundwater table (cm-mv)

glg

Upper value for groundwater table (cm-mv)

B_GWL_CLASS

grondwatertrap

nf

Original values of N run-off fraction to surface water (kg N drain/ha/year per kg N overschot/ha/year) or groundwater (mg NO3/L per kg N overschot/ha/year)

leaching_to-set

Tells if leaching to ground water or surface water)