Calculate the Open Bodem Index score for a data table

This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI). In contrast to obic_field, this wrapper can handle a data.table as input. Multiple sites (distinguished in the column 'ID') can be simulated simultaneously.

obic_field_dt(dt, output = "all")

Arguments

dt

(data.table) A data.table containing the data of the fields to calculate the OBI

output

(character) An optional argument to select output: obic_score, scores, indicators, recommendations, or all. (default = all)