Effects of measures on soil indicators

This table defines the effects of 22 measures on soil indicators

recom.obic_bkp

Format

A data.frame with 9152 rows and 11 columns:

m_nr

The ID number of measure

m_description

The description of measure

m_prio

weighing factor for measure. This is not used in the script.

m_treshold

Threshold value of the indicator value. This is not used in the script.

m_order

Order of measures. When scores are tie, the measure with a smaller number is chosen.

m_soilfunction

description of the OBIC indicator variable

indicator

Name of OBIC soil indicator variable

m_effect

Effect of measure on soil indicator. 3/2/1/0/-1

m_sector

type of agricultural sector: dairy/arable/vegetable/tree cultivation (in dutch)

m_soiltype

type of soil: sand/clay/peat/loess (in dutch)

m_applicability

is the measure applicable for combination of sector and soil (1/0)