Linking table between crops, soils, groundwater tables and water induced stresses in OBIC

This table helps to link the different crops in the OBIC functions with the crops selected by the user

waterstress.obic

Format

A data.frame with x rows and y columns:

cropname

The name of the crop

soilunit

The category for this soil, derived from 1:50.000 soil map

gt

The class describing mean highest and lowest groundwater table, derived from 1:50.000 soil map

droughtstress

The mean yield reduction due to drought (in percentage)

wetnessstress

The mean yield reduction due to water surplus (in percentage)

waterstress

The mean combined effect water stress (due to deficiency or excess of water)