Linking table between crops and different functions in OBIC

This table helps to link the different crops in the OBIC functions with the crops selected by the user

crops.obic

Format

An object of class data.table (inherits from data.frame) with 525 rows and 22 columns.

Details

crop_code

The BRP gewascode of the crop

crop_name

The name of the crop, in lower case

crop_waterstress

Classification linking for linking crops to waterstress.obic

crop_intensity

Whether crop is root/tuber crop, rest crop, or other.

crop_eos

Effective soil organic matter produced by the crop in kg/ha

crop_eos_residue

Effective soil organic matter from plant residues in kg/ha

crop_category

Classification of crop per land use type (arable, maize, grass, nature)

crop_rotation

Classification of crop to determine function within crop rotations

crop_crumbleability

The category for this crop at crumbleability

crop_phosphate

The category for this crop for evaluation phosphate availability

crop_sealing

The category for this crop at soil sealing

crop_n

The category for this crop for evaluation nitrogen

crop_k

The category for this crop for evaluation potassium

crop_measure

The category for this crop for evaluating measures

nf_clay

Allowed effective N dose on clay soils

nf_sand.other

Allowed effective N dose on sandy soils

nf_sand.south

Allowed effective N dose on sandy soils sensitive to leaching

nf_loess

Allowed effective N dose on loess soils

nf_peat

Allowed effective N dose on peat soils

crop_name_scientific

All-lower-case scientific name of the crop species. When crop is not species specific the genus of the crop is given

crop_season

Crop category for length growing season

crop_makkink

Crop category for makkink correction factors