Calculate the Open Bodem Index score for one field

This functions wraps the functions of the OBIC into one main function to calculate the score for Open Bodem Index (OBI) for a single field.

obic_field(
  B_SOILTYPE_AGR,
  B_GWL_CLASS,
  B_SC_WENR,
  B_HELP_WENR,
  B_AER_CBS,
  B_LU_BRP,
  A_SOM_LOI,
  A_SAND_MI,
  A_SILT_MI,
  A_CLAY_MI,
  A_PH_CC,
  A_N_RT,
  A_CN_FR,
  A_S_RT,
  A_N_PMN,
  A_P_AL,
  A_P_CC,
  A_P_WA,
  A_CEC_CO,
  A_CA_CO_PO,
  A_MG_CO_PO,
  A_K_CO_PO,
  A_K_CC,
  A_MG_CC,
  A_MN_CC,
  A_ZN_CC,
  A_CU_CC,
  A_C_BCS = NA,
  A_CC_BCS = NA,
  A_GS_BCS = NA,
  A_P_BCS = NA,
  A_RD_BCS = NA,
  A_EW_BCS = NA,
  A_SS_BCS = NA,
  A_RT_BCS = NA,
  A_SC_BCS = NA,
  M_COMPOST = NA_real_,
  M_GREEN = NA,
  M_NONBARE = NA,
  M_EARLYCROP = NA,
  M_SLEEPHOSE = NA,
  M_DRAIN = NA,
  M_DITCH = NA,
  M_UNDERSEED = NA,
  M_LIME = NA,
  M_NONINVTILL = NA,
  M_SSPM = NA,
  M_SOLIDMANURE = NA,
  M_STRAWRESIDUE = NA,
  M_MECHWEEDS = NA,
  M_PESTICIDES_DST = NA,
  ID = 1,
  output = "all"
)

Arguments

B_SOILTYPE_AGR
(character) The agricultural type of soil

B_GWL_CLASS
(character) The groundwater table class

B_SC_WENR
(character) The risk for subsoil compaction as derived from risk assessment study of Van den Akker (2006).

B_HELP_WENR
(character) The soil type abbreviation, derived from 1:50.000 soil map

B_AER_CBS
(character) The agricultural economic region in the Netherlands (CBS, 2016)

B_LU_BRP
(numeric) a series with crop codes given the crop rotation plan (source: the BRP)

A_SOM_LOI
(numeric) The percentage organic matter in the soil (%)

A_SAND_MI
(numeric) The sand content of the soil (%)

A_SILT_MI
(numeric) The silt content of the soil (%)

A_CLAY_MI
(numeric) The clay content of the soil (%)

A_PH_CC
(numeric) The acidity of the soil, measured in 0.01M CaCl2 (-)

A_N_RT
(numeric) The organic nitrogen content of the soil in mg N / kg

A_CN_FR
(numeric) The carbon to nitrogen ratio (-)

A_S_RT
(numeric) The total Sulpher content of the soil (in mg S per kg)

A_N_PMN
(numeric) The potentially mineralizable N pool (mg N / kg soil)

A_P_AL
(numeric) The P-AL content of the soil

A_P_CC
(numeric) The plant available P content, extracted with 0.01M CaCl2 (mg / kg)

A_P_WA
(numeric) The P-content of the soil extracted with water (mg P2O5 / 100 ml soil)

A_CEC_CO
(numeric) The cation exchange capacity of the soil (mmol+ / kg), analysed via Cobalt-hexamine extraction

A_CA_CO_PO
(numeric) The The occupation of the CEC with Ca (%)

A_MG_CO_PO
(numeric) The The occupation of the CEC with Mg (%)

A_K_CO_PO
(numeric) The occupation of the CEC with K (%)

A_K_CC
(numeric) The plant available K content, extracted with 0.01M CaCl2 (mg / kg)

A_MG_CC
(numeric) The plant available Mg content, extracted with 0.01M CaCl2 (ug / kg)

A_MN_CC
(numeric) The plant available Mn content, extracted with 0.01M CaCl2 (ug / kg)

A_ZN_CC
(numeric) The plant available Zn content, extracted with 0.01M CaCl2 (ug / kg)

A_CU_CC
(numeric) The plant available Cu content, extracted with 0.01M CaCl2 (ug / kg)

A_C_BCS
(numeric) The presence of visible cracks in the top layer (optional, score 0-1-2)

A_CC_BCS
(integer) The crop cover on the surface (optional, score 0-1-2)

A_GS_BCS
(numeric) The presence of waterlogged conditions, gley spots (optional, score 0-1-2)

A_P_BCS
(numeric) The presence / occurrence of water puddles on the land, ponding (optional, score 0-1-2)

A_RD_BCS
(integer) The rooting depth (optional, score 0-1-2)

A_EW_BCS
(numeric) The presence of earth worms (optional, score 0-1-2)

A_SS_BCS
(integer) The soil structure (optional, score 0-1-2)

A_RT_BCS
(numeric) The presence of visible tracks / rutting or trampling on the land (optional, score 0-1-2)

A_SC_BCS
(numeric) The presence of compaction of subsoil (optional, score 0-1-2)

M_COMPOST
(numeric) The frequency that compost is applied (optional, every x years)

M_GREEN
(boolean) A soil measure. Are catch crops sown after main crop (optional, option: yes or no)

M_NONBARE
(boolean) A soil measure. Is parcel for 80 percent of the year cultivated and 'green' (optional, option: yes or no)

M_EARLYCROP
(boolean) A soil measure. Use of early crop varieties to avoid late harvesting (optional, option: yes or no)

M_SLEEPHOSE
(boolean) A soil measure. Is sleephose used for slurry application (optional, option: yes or no)

M_DRAIN
(boolean) A soil measure. Are under water drains installed in peaty soils (optional, option: yes or no)

M_DITCH
(boolean) A soil measure. Are ditched maintained carefully and slib applied on the land (optional, option: yes or no)

M_UNDERSEED
(boolean) A soil measure. Is grass used as second crop in between maize rows (optional, option: yes or no)

M_LIME
(boolean) measure. Has field been limed in last three years (option: yes or no)

M_NONINVTILL
(boolean) measure. Non inversion tillage (option: yes or no)

M_SSPM
(boolean) measure. Soil Structure Protection Measures, such as fixed driving lines, low pressure tires, and light weighted machinery (option: yes or no)

M_SOLIDMANURE
(boolean) measure. Use of solid manure (option: yes or no)

M_STRAWRESIDUE
(boolean) measure. Application of straw residues (option: yes or no)

M_MECHWEEDS
(boolean) measure. Use of mechanical weed protection (option: yes or no)

M_PESTICIDES_DST
(boolean) measure. Use of DST for pesticides (option: yes or no)

ID
(character) A field id

output
(character) An optional argument to select output: obic_score, scores, indicators, recommendations, or all. (default = all)